Feature: Clean up code
