+++ constants/popularAuthors.ts
@@ -1,0 +1,6 @@
+export const POPULAR_AUTHORS = [
+  "/authors/OL23919A",
+  "/authors/OL26320A",
+  "/authors/OL2622837A",
+  "/authors/OL24638A"
+];
