--- screens/authentication/Settings/SettingsScreen.tsx
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function SettingsScreen() {
-  return (
-    <View>
-      <Text>Settings Screen</Text>
-    </View>
-  );
-}
