+++ types/IAuthor.ts
@@ -1,0 +1,4 @@
+export interface Author {
+  key: string;
+  name: string;
+}
