+++ types/IAuthorWork.ts
@@ -1,0 +1,4 @@
+export interface AuthorWork {
+  key: string;
+  title: string;
+}
