+++ types/IBookDetailsResponse.ts
@@ -1,0 +1,4 @@
+export interface IBookDetailsResponse {
+  description?: string | { value: string };
+  subjects?: string[];
+}
