+++ types/navigation.types.ts
@@ -1,0 +1,8 @@
+export type RootStackParamList = {
+  Login: undefined;
+  Home: undefined;
+  Register: undefined;
+  BookList: undefined;
+  Search: undefined;
+  Account: undefined;
+};
